requirements.txt
torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1
opencv-python==4.10.0.84
mediapipe==0.10.14
numpy==1.26.4
pillow==10.3.0
tqdm==4.66.5
pyttsx3==2.90
gTTS==2.5.1
playsound==1.3.0
matplotlib==3.9.2
scikit-learn==1.5.2
pandas==2.2.3
